Search Engine implementation
A search engine must be able to handle the high demand and it also needs to deliver useful results to the users. Since there are large number of web pages, an efficient data flow for accurate results must be considered. The aim of this project is to create a search engine using the concepts that we have discussed in previous projects.